server/routes refactor
