+++ server/routers/__init__.py
@@ -1,0 +1,1 @@
+# Routers package for FastAPI modular route groupings.
